requirements.txt
fastapi
uvicorn
pydantic
requests
beautifulsoup4
pytesseract
Pillow
openai
faiss-cpu
python-dotenv
